[snack-sdk] webplayer transport should not rely on pubnub
Summary:
    Original webplayer transport still leverages PubNub as underlying
    transport, this commit changes to pure window.postMessage
    implementation.

    This commit also removes the unused `timedout` payload from PubNub.
